+++ worker/src/test/java/gov/cms/ab2d/worker/service/JobProcessingServiceTest.java
@@ -1,0 +1,141 @@
+package gov.cms.ab2d.worker.service;
+
+import gov.cms.ab2d.common.model.Contract;
+import gov.cms.ab2d.common.model.Job;
+import gov.cms.ab2d.common.model.JobStatus;
+import gov.cms.ab2d.common.model.Sponsor;
+import gov.cms.ab2d.common.model.User;
+import gov.cms.ab2d.common.repository.ContractRepository;
+import gov.cms.ab2d.common.repository.JobRepository;
+import gov.cms.ab2d.common.repository.SponsorRepository;
+import gov.cms.ab2d.common.repository.UserRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import java.io.IOException;
+import java.time.OffsetDateTime;
+import java.util.Random;
+
+import static java.lang.Boolean.TRUE;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.junit.Assert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+@SpringBootTest
+class JobProcessingServiceTest {
+    private Random random = new Random();
+
+    @Autowired
+    private JobProcessingService cut;
+
+    @Autowired
+    private JobRepository jobRepository;
+    @Autowired
+    private SponsorRepository sponsorRepository;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private ContractRepository contractRepository;
+
+    private Sponsor sponsor;
+    private User user;
+    private Job job;
+
+    @BeforeEach
+    void setUp() {
+        jobRepository.deleteAll();
+        userRepository.deleteAll();
+        sponsorRepository.deleteAll();
+
+        sponsor = createSponsor();
+        user = createUser(sponsor);
+        job = createJob(user);
+    }
+
+
+    @Test
+    @DisplayName("When a job is in submitted status, it can be put into progress upon starting processing")
+    void whenJobIsInSubmittedStatus_ThenJobShouldBePutInProgress() {
+        var processedJob = cut.putJobInProgress("S001");
+        assertThat(processedJob.getStatus(), is(JobStatus.IN_PROGRESS));
+    }
+
+    @Test
+    @DisplayName("When a job is not already in a submitted status, it cannot be put into progress")
+    void whenJobIsNotInSubmittedStatus_ThenJobShouldNotBePutInProgress() {
+        job.setStatus(JobStatus.IN_PROGRESS);
+        jobRepository.save(job);
+
+        var exceptionThrown = assertThrows(IllegalArgumentException.class,
+                () -> cut.putJobInProgress("S001"));
+
+        assertThat(exceptionThrown.getMessage(), is("Job S001 is not in SUBMITTED status."));
+    }
+
+
+    @Test
+    @DisplayName("When a job is in submitted status, it can be processed")
+    void processJob() throws IOException {
+
+        job.setStatus(JobStatus.IN_PROGRESS);
+        jobRepository.save(job);
+        var contract = createContract(sponsor);
+
+        var processedJob = cut.processJob("S001");
+
+        assertThat(processedJob.getStatus(), is(JobStatus.SUCCESSFUL));
+        assertThat(processedJob.getStatusMessage(), is("100%"));
+        assertThat(processedJob.getExpiresAt(), notNullValue());
+    }
+
+    private Sponsor createSponsor() {
+        Sponsor sponsor = new Sponsor();
+        sponsor.setId(Long.valueOf(random.nextInt()));
+        sponsor.setOrgName("Hogwarts School of Wizardry");
+        sponsor.setLegalName("Hogwarts School of Wizardry LLC");
+        sponsor.setHpmsId(random.nextInt());
+        return sponsorRepository.save(sponsor);
+    }
+
+    private User createUser(Sponsor sponsor) {
+        User user = new User();
+        user.setId(Long.valueOf(random.nextInt()));
+        user.setUserName("Harry_Potter");
+        user.setFirstName("Harry");
+        user.setLastName("Potter");
+        user.setEmail("harry_potter@hogwarts.com");
+        user.setEnabled(TRUE);
+        user.setSponsor(sponsor);
+        return userRepository.save(user);
+    }
+
+    private Contract createContract(Sponsor sponsor) {
+        final int anInt = random.nextInt(299);
+        Contract contract = new Contract();
+        contract.setId(Long.valueOf(anInt));
+        contract.setContractName("CONTRACT_0000" + anInt);
+        contract.setContractNumber("CONTRACT_0000" + anInt);
+        contract.setAttestedOn(OffsetDateTime.now().minusDays(10));
+        contract.setSponsor(sponsor);
+
+        sponsor.getContracts().add(contract);
+        return contractRepository.save(contract);
+    }
+
+    private Job createJob(User user) {
+        Job job = new Job();
+        job.setId(Long.valueOf(random.nextInt()));
+        job.setJobUuid("S001");
+        job.setStatus(JobStatus.SUBMITTED);
+        job.setStatusMessage("0%");
+        job.setUser(user);
+        job.setCreatedAt(OffsetDateTime.now());
+        return jobRepository.save(job);
+    }
+
+
+}
